% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area Under the Curve (histogram)}
\usage{
auc(u = NULL, x = NULL, y = NULL, nbins = 30)
}
\arguments{
\item{u}{Numeric vector of samples supported on unit interval with which to
create a histogram (use \code{u = NULL} if \code{x} and \code{y} are supplied).}

\item{x}{Numeric vector of histogram locations. (Not used if \code{u} is supplied).}

\item{y}{Numeric vector of histogram heights OR function evaluating the curve
for a given value of \code{u} supported on (0,1). (Not used if \code{u} is supplied).}

\item{nbins}{Number of histogram bins to use (defaults to 30).}
}
\description{
Function auc() is deprecated. Use utility_shrinkslice(..., utility_type = "AUC").
}
