% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapproxt.R
\name{lapproxt}
\alias{lapproxt}
\title{Cauchy Pseudo-Target from Laplace Approximation}
\usage{
lapproxt(lf, init, sc_adj = 1, lb = -Inf, ub = Inf, maxit = 100, ...)
}
\arguments{
\item{lf}{Univariate function evaluating the unnormalized log density to approximate.}

\item{init}{Numeric scalar for an initial value (used in optimization).}

\item{sc_adj}{Positive numeric scalar; manual multiplicative adjustment to the
scale of the output pseudo-target.}

\item{lb}{Numeric scalar giving the value of left truncation of the resulting pseudo-target.
Defaults to \code{-Inf}.}

\item{ub}{Numeric scalar giving the value of right truncation of the resulting pseudo-target.
Defaults to \code{Inf}.}

\item{maxit}{See \link[stats]{optim}.}

\item{...}{See \link[stats]{optim}.}
}
\description{
Author: Sam Johnson
}
\examples{
pseu <- lapproxt(function(x) dnorm(x, log = TRUE), init = 0.5, lb = -1.0)
curve(dnorm(x)/(1- pnorm(-1)), from = -1, to = 6, col = "blue")
curve(pseu$d(x), from = -2, to = 6, add = TRUE)
}
