% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_slice_width.R
\name{meanSliceWidth_int}
\alias{meanSliceWidth_int}
\title{Expected Slice Width (integration)}
\usage{
meanSliceWidth_int(h, tol = 0.005)
}
\arguments{
\item{h}{Function evaluating an unnormalized density supported on the unit interval.}

\item{tol}{Numerical scalar passed to \code{abs.tol} in the call to \code{integrate()}.}
}
\description{
Calculate the expected slice width of a slice sampler for an unnormalized
density supported on the unit interval.
}
\details{
Uses numeric integration to evaluate a double integral.
}
