% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_targets.R
\name{pseudo_beta_list}
\alias{pseudo_beta_list}
\title{Specify a beta Pseudo-Target}
\usage{
pseudo_beta_list(shape1, shape2, log_p = FALSE, name = NULL)
}
\arguments{
\item{shape1}{Positive numeric scalar giving the first shape parameter.}

\item{shape2}{Positive numeric scalar giving the second shape parameter.}

\item{log_p}{(Not implemented) Logical: evaluate distribution and quantile functions using the log probability.}

\item{name}{String appending optional message to the textual name of the distribution.}
}
\value{
A list with named components:

\code{d}: function to evaluate the density

\code{ld}: function to evaluate the log density

\code{q}: function to evaluate the quantile function

\code{p}: function to evaluate the distribution function

\code{t}: text description of the distribution

\code{params}: returns the parameters passed to the function
}
\description{
Create a list of functions to evaluate a beta pseudo-target
with supplied shape parameters.
}
\examples{
pseu <- pseudo_beta_list(shape1 = 2.0, shape2 = 1.0)
str(pseu)
pseu$d(0.5)
pseu$ld(0.5)
pseu$p(0.5)
pseu$q(0.25)
}
