% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperrectangle_slice.R
\name{pseudo_t_condseq}
\alias{pseudo_t_condseq}
\title{Sequence of Conditional Pseudo-Targets from a Realization}
\usage{
pseudo_t_condseq(x, pseu_init, loc_fn, sc_fn, degf, lb, ub)
}
\arguments{
\item{x}{A numeric vector of values between 0 and 1.}

\item{pseu_init}{A list output from \code{pseu_t_list()} describing the
marginal pseudo-target for \code{x[1]}.}

\item{loc_fn}{A function that specifies the location of a conditional
pseudo-target given the elements in \code{x} that precede it.}

\item{sc_fn}{A function that specifies the scale of a conditional
pseudo-target given the elements in \code{x} that precede it}

\item{degf}{A positive scalar for the degrees of freedom for all conditional
pseudo-targets.}

\item{lb}{A numeric vector (same length as \code{x}) specifying the lower
bound of support for each conditional pseudo-target.}

\item{ub}{A numeric vector (same length as \code{x}) specifying the upper
bound of support for each conditional pseudo-target.}
}
\value{
A list containing \code{x} obtained from the sequence of inverse
cdfs, and \code{pseudo_t_seq}, a list output from \code{pseu_t_list()}
describing the sequence of conditional pseudo-targets.
}
\description{
Given a realization of a random vector, generate a the corresponding
sequence of conditional pseudo-target inverse cdfs.
The pseudo-target is specified as
a sequence of growing conditional distributions.
}
