% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_targets.R
\name{pseudo_t_list}
\alias{pseudo_t_list}
\title{Specify a Pseudo-Target within the Student-t Class}
\usage{
pseudo_t_list(loc, sc, degf, lb = -Inf, ub = Inf, log_p = FALSE, name = NULL)
}
\arguments{
\item{loc}{Numeric scalar giving the location parameter.}

\item{sc}{Positive numeric scalar giving the scale parameter.}

\item{degf}{Positive numeric scalar giving the degrees of freedom parameter.}

\item{lb}{Numeric scalar giving the value of left truncation. Defaults to \code{-Inf}.}

\item{ub}{Numeric scalar giving the value of right truncation. Defaults to \code{Inf}.}

\item{log_p}{(Not implemented) Logical: evaluate distribution and quantile functions using the log probability.}

\item{name}{String appending optional message to the textual name of the distribution.}
}
\value{
A list with named components:

\code{d}: function to evaluate the density

\code{ld}: function to evaluate the log density

\code{q}: function to evaluate the quantile function

\code{p}: function to evaluate the distribution function

\code{txt}: text description of the distribution

\code{params}: returns the parameters passed to the function

\code{lb}: lower boundary of support

\code{ub}: upper boundary of support
}
\description{
Create a list of functions to evaluate a pseudo-target in the Student-t class
with supplied location, scale, and degrees of freedom. The distribution is optionally
truncated to specified bounds (and renormalized).
}
\examples{
pseu <- pseudo_t_list(loc = 0.0, sc = 1.0, degf = 5.0, lb = 0.0, ub = Inf)
str(pseu)
pseu$d(1.5)
pseu$ld(1.5)
pseu$p(1.5)
pseu$q(0.8060963)
}
