% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_samplers_univariate.R
\name{slice_generalized_elliptical}
\alias{slice_generalized_elliptical}
\title{General Elliptical Slice Sampler (Univariate)}
\usage{
slice_generalized_elliptical(x, target, mu, sigma, df)
}
\arguments{
\item{x}{The current state (as a numeric scalar).}

\item{target}{A function taking numeric scalar that evaluates the log-target
density, returning a numeric scalar.}

\item{mu}{A numeric scalar with the mean of the normal to be sampled}

\item{sigma}{A numeric scalar with the standard deviation of the normal to be sampled.}

\item{df}{Degrees of freedom of Student t pseudo-target.}
}
\value{
A list contains two elements: "x" is the new state and "nEvaluations"
is the number of evaluations of the target function used to obtain the new
state.
}
\description{
General Elliptical Slice Sampler of Nishihara (2014)
}
\examples{
lf <- function(x) dbeta(x, 3, 4, log = TRUE)
draws <- numeric(10) # set to numeric(1e3) for more complete illustration
nEvaluations <- 0L
for (i in seq.int(2, length(draws))) {
  out <- slice_generalized_elliptical(draws[i - 1], target = lf,
                                      mu = 0.5, sigma = 1, df = 5)
  draws[i] <- out$x
  nEvaluations <- nEvaluations + out$nEvaluations
}
nEvaluations / length(draws)
plot(density(draws), xlim = c(0, 1))
curve(exp(lf(x)), 0, 1, col = "blue", add = TRUE)

}
