% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_shrinkslice.R
\name{utility_shrinkslice}
\alias{utility_shrinkslice}
\title{Utility for a Transformed Target}
\usage{
utility_shrinkslice(
  h = NULL,
  x = NULL,
  u = NULL,
  type = "samples",
  nbins = 30,
  plot = FALSE,
  utility_type = "AUC",
  options = list(tol_int = 0.001, n_opt = 10, interval_opt = c(1e-06, 0.999999))
)
}
\arguments{
\item{h}{Function to evaluate the unnormalized transformed target
\eqn{h(\psi) = g(\hat{\Pi}^{-1}(\psi))/\hat{\pi}(\hat{\Pi}^{-1}(\psi))}
with argument \eqn{\psi \in (0,1)}. Use \code{h = NULL} if type = "samples".}

\item{x}{Numeric vector of histogram locations. Not used if \code{u} is supplied or type = "integration". If used but not supplied, a default grid is chosen.}

\item{u}{Numeric vector of samples supported on unit interval (\eqn{\psi}) with which to
create histogram (use \code{u = NULL} if type = "integration" or "grid").}

\item{type}{String specifying the utility calculation method. One of "samples", "integration", or "grid".
Use of "integration" requires \code{h}. Use of "samples" requires \code{u}.
Use of "grid" requires \code{h} and optionally \code{x}. Default is to use "samples".}

\item{nbins}{Number of histogram bins to use (defaults to 30). Overridden by the length of x.}

\item{plot}{Logical for whether to plot a visualization of the transformed target.
Defaults to \code{FALSE}.}

\item{utility_type}{String identifying utility type, either "AUC" (default) or "MSW"}

\item{options}{List containing the following. \code{tol_int}: Positive numeric scalar that passes to \code{abs.tol} in the call to \code{integrate()} used for type = "integration".
\code{n_opt}: Positive integer giving the number of initialization points for numerical optimization in AUC with type = "integration".
\code{interval_opt}: Numerical vector of length two giving the interval for numerical optimization in AUC with type = "integration".}
}
\value{
Scalar value of the utility function evaluation.
}
\description{
Evaluates the utility function for a transformed target, which can be one of
Area Under the Curve (AUC) and Mean Slice Width (MSW).
}
\details{
Calculating the utility with a function and \code{type = 'integration'} employs numerical optimization (for AUC) and numerical integration (for both AUC and MSW), which may not succeed for highly multimodal or narrow target densities.
}
\examples{
a <- 4
b <- 2
u <- rbeta(10e3, a, b)
h <- function(x) dbeta(x, a, b)
x <- sort(runif(20))
type <- "AUC"
type <- "MSW"

utility_shrinkslice(
  h = h,
  x = NULL,
  u = NULL,
  type = "integration",
  plot = TRUE,
  utility_type = type
)

utility_shrinkslice(
  h = h,
  x = NULL,
  u = NULL,
  nbins = 50,
  type = "grid",
  plot = TRUE,
  utility_type = type
)

utility_shrinkslice(
  h = h,
  x = x,
  u = NULL,
  type = "grid",
  plot = TRUE,
  utility_type = type
)

utility_shrinkslice(
  h = NULL,
  x = NULL,
  u = u,
  nbins = 30,
  type = "samples",
  plot = TRUE,
  utility_type = type
)

utility_shrinkslice(
  h = NULL,
  x = x,
  u = u,
  type = "samples",
  plot = TRUE,
  utility_type = type
)

}
\references{
Heiner, M. J., Johnson, S. B., Christensen, J. R., and Dahl, D. B. (2026+), "Quantile Slice Sampling," \emph{arXiv preprint arXiv:2407.12608} \doi{https://doi.org/10.48550/arXiv.2407.12608}
}
