% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperrectangle_slice.R
\name{pseudo_condseq_XfromU}
\alias{pseudo_condseq_XfromU}
\title{Inverse Transform from Sequence of Conditional Pseudo-Targets}
\usage{
pseudo_condseq_XfromU(u, pseudo_init, loc_fn, sc_fn, lb, ub)
}
\arguments{
\item{u}{A numeric vector of values between 0 and 1.}

\item{pseudo_init}{A list output from \code{pseudo_list()} describing the
marginal pseudo-target for \code{x[1]}.}

\item{loc_fn}{A function that specifies the location of a conditional
pseudo-target given the elements in \code{x} that precede it.}

\item{sc_fn}{A function that specifies the scale of a conditional
pseudo-target given the elements in \code{x} that precede it}

\item{lb}{A numeric vector (same length as \code{x}) specifying the lower
bound of support for each conditional pseudo-target.}

\item{ub}{A numeric vector (same length as \code{x}) specifying the upper
bound of support for each conditional pseudo-target.}
}
\value{
A list containing \code{x} obtained from the sequence of inverse
cdfs, and \code{pseudo_seq}, a list output from \code{pseu_list()}
describing the sequence of conditional pseudo-targets.
}
\description{
Given a vector of from a unit hypercube, map to the original vector using
a sequence of conditional pseudo-target inverse cdfs.
The pseudo-target is specified as
a sequence of growing conditional distributions.
}
\details{
See the documentation for \code{slice_mv_transform_seq()} for examples.
}
