% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical_slice_multivariate.R
\name{slice_elliptical_mv}
\alias{slice_elliptical_mv}
\title{Multivariate Elliptical Slice Sampler}
\usage{
slice_elliptical_mv(x, target, mu, Sig, is_chol = FALSE)
}
\arguments{
\item{x}{The current state (as a numeric scalar).}

\item{target}{A function taking numeric scalar that evaluates the log-target
density, returning a numeric scalar.}

\item{mu}{Numeric vector with the mean of the normal to be sampled.}

\item{Sig}{Positive definite covariance matrix. Alternatively, a
lower-triangular matrix with the Cholesky factor of the covariance matrix
(for faster computation).}

\item{is_chol}{Logical, is the supplied \code{Sig} in Cholesky (lower triangular) format?}
}
\value{
A list contains two elements: "x" is the new state and "nEvaluations"
is the number of evaluations of the target function used to obtain the new
state.
}
\description{
Algorithm 1 of Nishihara et al (2014) of the
elliptical slice sampler of Murray, Adams, MacKay (2010).
}
\examples{
lf <- function(x) dbeta(x[1], 3, 4, log = TRUE) + dbeta(x[2], 5, 3, log = TRUE)
draws <- matrix(0.3, nrow = 1000, ncol = 2)
nEvaluations <- 0L
for (i in seq.int(2, nrow(draws))) {
  out <- slice_elliptical_mv(draws[i - 1,], target = lf,
              mu = c(0.5, 0.5), Sig = matrix(c(0.5, 0.25, 0.25, 0.5), nrow = 2))
  draws[i,] <- out$x
  nEvaluations <- nEvaluations + out$nEvaluations
}
nEvaluations / nrow(draws)
plot(draws[,1], draws[,2], xlim = c(0, 1))

}
