% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical_slice_multivariate.R
\name{slice_genelliptical_mv}
\alias{slice_genelliptical_mv}
\title{General Elliptical Slice Sampler (Multivariate)}
\usage{
slice_genelliptical_mv(x, target, mu, Sig, df, is_chol = FALSE)
}
\arguments{
\item{x}{The current state (as a numeric scalar).}

\item{target}{A function taking numeric scalar that evaluates the log-target
density, returning a numeric scalar.}

\item{mu}{Numeric vector with the mean of the normal to be sampled.}

\item{Sig}{Positive definite covariance matrix. Alternatively, a
lower-triangular matrix with the Cholesky factor of the covariance matrix
(for faster computation).}

\item{df}{Degrees of freedom of Student t pseudo-target.}

\item{is_chol}{Logical, is the supplied \code{Sig} in Cholesky (lower triangular) format?}
}
\value{
A list contains two elements: "x" is the new state and "nEvaluations"
is the number of evaluations of the target function used to obtain the new
state.
}
\description{
General Elliptical Slice Sampler, Algorithm 2 of Nishihara (2014)
}
\examples{
lf <- function(x) dbeta(x[1], 3, 4, log = TRUE) + dbeta(x[2], 5, 3, log = TRUE)
draws <- matrix(0.3, nrow = 10e3, ncol = 2)
nEvaluations <- 0L
for (i in seq.int(2, nrow(draws))) {
  out <- slice_genelliptical_mv(draws[i - 1,], target = lf,
              mu = c(0.5, 0.5), Sig = matrix(c(0.5, 0.25, 0.25, 0.5), nrow = 2),
              df = 5)
  draws[i,] <- out$x
  nEvaluations <- nEvaluations + out$nEvaluations
}
nEvaluations / nrow(draws)
nEvaluations / ess(draws[,1])
plot(draws[,1], draws[,2], xlim = c(0, 1))
hist(draws[,1], freq = FALSE); curve(dbeta(x, 3, 4), col = "blue", add = TRUE)
hist(draws[,2], freq = FALSE); curve(dbeta(x, 5, 3), col = "blue", add = TRUE)

}
