% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_samplers_univariate.R
\name{slice_latent}
\alias{slice_latent}
\title{Latent Slice Sampler}
\usage{
slice_latent(x, s, target, rate)
}
\arguments{
\item{x}{The current state (as a numeric scalar).}

\item{s}{A random variable that determines how far the algorithm samples from
on each side}

\item{target}{A function taking numeric scalar that evaluates the log-target
density, returning a numeric scalar.}

\item{rate}{The rate parameter for a truncated exponential.}
}
\value{
A list contains three elements: "x" is the new state of the target variable,
"s" is the new state of the latent scale variable, and "nEvaluations"
is the number of evaluations of the target function used to obtain the new
state.
}
\description{
Latent slice sampler of Li and Walker (2020).
}
\examples{
lf <- function(x) dbeta(x, 3, 4, log = TRUE)
draws <- numeric(10) # set to numeric(1e3) for more complete illustration
nEvaluations <- 0L
s <- 0.5
for (i in seq.int(2, length(draws))) {
  out <- slice_latent(draws[i - 1], s, target = lf, rate = 0.3)
  draws[i] <- out$x
  s <- out$s
  nEvaluations <- nEvaluations + out$nEvaluations
}
nEvaluations / length(draws)
plot(density(draws), xlim = c(0, 1))
curve(exp(lf(x)), 0, 1, col = "blue", add = TRUE)

}
