% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_shrinkslice.R
\name{utility_shrinkslice}
\alias{utility_shrinkslice}
\title{Utility for a Transformed Target}
\usage{
utility_shrinkslice(
  h = NULL,
  x = NULL,
  y = NULL,
  u = NULL,
  type = "samples",
  nbins = 30,
  plot = FALSE,
  utility_type = "AUC",
  tol_int = 0.001
)
}
\arguments{
\item{h}{Function to evaluate the unnormalized transformed target
\eqn{h(\psi) = g(\hat{\Pi}^{-1}(\psi))/\hat{\pi}(\hat{\Pi}^{-1}(\psi))}
with argument \eqn{\psi \in (0,1)}.}

\item{x}{Numeric vector of histogram locations. (Not used if \code{u} is supplied).}

\item{y}{Numeric vector of histogram heights OR function evaluating the curve
for a given value of \code{u}.(Not used if \code{u} is supplied).}

\item{u}{Numeric vector of samples supported on unit interval (\eqn{\psi}) with which to
create histogram (use \code{u = NULL} if \code{x} and \code{y} are supplied).}

\item{type}{String specifying the input type. One of "function", "samples", or "grid".
Use of "function" requires \code{h}. Use of "samples" requires \code{u}.
Use of "grid" requires \code{h} and optionally \code{x}. Default is to use "samples".}

\item{nbins}{Number of histogram bins to use (defaults to 30).}

\item{plot}{Logical for whether to plot a visualization of the transformed target.
Defaults to \code{FALSE}.}

\item{utility_type}{String identifying utility type, either AUC (default) or MSW}

\item{tol_int}{Positive numeric scalar that passes to \code{abs.tol} in the call to \code{integrate()}.
Defaults to \code{1.0e-3}.}
}
\value{
Scalar value of the utility function evaluation.
}
\description{
Evaluates the utility function for a transformed target, which can be one of
Area Under the Curve (AUC) and Mean Slice Width (MSW).
}
\examples{
x <- seq(0.001, 0.999, length = 1000)
y <- dbeta(x, 2.0, 2.0)
utility_shrinkslice(x = x, y = y, type = "grid", plot = TRUE, use_meanSliceWidth = TRUE)
utility_shrinkslice(h = \(x) dbeta(x, 2.0, 2.0), type = "function", plot = TRUE,
                    use_meanSliceWidth = TRUE)
utility_shrinkslice(u = rbeta(1e3, 2.0, 2.0), type = "samples")
}
